fix: 멤버 id를 조회하는 방식으로 변경
